% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_discharge_conversions.r
\name{cfs_to_cms}
\alias{cfs_to_cms}
\title{cubic feet/second to cubic meters/second}
\usage{
cfs_to_cms(cfs)
}
\arguments{
\item{cfs}{discharge (cubic feet per second)}
}
\value{
discharge (cubic meters per second)
}
\description{
round to 3 significant figures, as that is precision of USGS flow gauges
}
\examples{
hfe <- cfs_to_cms(40000)
loadfollowing <- cfs_to_cms(12000, 16000)
}
