% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_discharge_conversions.r
\name{cms_to_cfs}
\alias{cms_to_cfs}
\title{cubic meters/second to cubic feet/second}
\usage{
cms_to_cfs(cms)
}
\arguments{
\item{cfs}{discharge (cubic feet per second)}
}
\value{
discharge (cubic meters per second)
}
\description{
round to 3 significant figures, as that is precision of USGS flow gauges
}
\examples{
hfe <- cfs_to_cms(35.2)
loadfollowing <- cfs_to_cms(12.3, 18.1)
}
