% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length_conversions.r
\name{fork_from_total}
\alias{fork_from_total}
\title{Predict Total Length from Fork Length}
\usage{
fork_from_total(species.code, total.length)
}
\arguments{
\item{species.code}{3 letter GCMRC species code (i.e., "FMS", "HBC", "RBT")}

\item{total.length}{fork length (mm)}
}
\value{
fork length (mm)
}
\description{
Predict Total Length from Fork Length
}
\examples{
fork_from_total("HBC", 282)
}
