% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length_conversions.r
\name{fork_to_total}
\alias{fork_to_total}
\title{Predict Total Length from Fork Length}
\usage{
fork_to_total(species.code, fork.length)
}
\arguments{
\item{species.code}{3 letter GCMRC species code (i.e., "FMS", "HBC", "RBT")}

\item{fork.length}{fork length (mm)}
}
\value{
total length (mm)
}
\description{
Predict Total Length from Fork Length
}
\examples{
fork_to_total("FMS", 437)
}
