% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_discharge_conversions.r
\name{km_to_mile}
\alias{km_to_mile}
\title{Convert kilometers to miles}
\usage{
km_to_mile(km)
}
\arguments{
\item{km}{kilometer}
}
\value{
mile
}
\description{
use for river km to mile conversion if you are not changing km/mile 0 location
}
\examples{
mile1 <- km_to_mile(3.42)
mile2 <- mile_to_km(c(150.23, 358.3))
}
