% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_discharge_conversions.r
\name{km_to_mile_COR}
\alias{km_to_mile_COR}
\title{Convert river kilometers (rkm 0 = Glen Canyon Dam) to miles (RM 0 = Lees Ferry)}
\usage{
km_to_mile_COR(rkm)
}
\arguments{
\item{rkm}{GCMRC river kilometer, numbered downstream where RM 0 = Glen Canyon Dam}
}
\value{
river mile, numbered downstream where RM 0 = Lees Ferry
}
\description{
Glen Canyon Dam is 15.8 miles (25.42 km) upstream of Lees Ferry
}
\examples{
mile1 <- km_to_mile_COR(3.42)
mile2 <- mile_to_km_COR(c(150.23, 358.3))
}
