% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_discharge_conversions.r
\name{mile_to_km}
\alias{mile_to_km}
\title{Convert  miles to kilometers}
\usage{
mile_to_km(mile)
}
\arguments{
\item{mile}{mile}
}
\value{
kilometer
}
\description{
use for river mile to km conversion if you are not changing km/mile 0 location
}
\examples{
four.mile.bar <- mile_to_km(-4.02)
confluences <- mile_to_km(c(61.4, 89.1, 157))
}
