% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length_conversions.r
\name{total_from_fork}
\alias{total_from_fork}
\title{Predict Total Length from Fork Length}
\usage{
total_from_fork(species.code, fork.length)
}
\arguments{
\item{species.code}{3 letter GCMRC species code (i.e., "FMS", "HBC", "RBT")}

\item{fork.length}{fork length (mm)}
}
\value{
total length (mm)
}
\description{
Predict Total Length from Fork Length
}
\examples{
total_from_fork("FMS", 437)
}
