% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length_conversions.r
\name{total_to_fork}
\alias{total_to_fork}
\title{Predict Fork Length from Total Length}
\usage{
total_to_fork(species.code, total.length)
}
\arguments{
\item{species.code}{3 letter GCMRC species code (i.e., "FMS", "HBC", "RBT")}

\item{total.length}{fork length (mm)}
}
\value{
fork length (mm)
}
\description{
Predict Fork Length from Total Length
}
\examples{
total_to_fork("HBC", 282)
}
