% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length_conversions.r
\name{weight_from_fork}
\alias{weight_from_fork}
\title{Predict Weight (g) from Fork Length (mm)}
\usage{
weight_from_fork(species.code, fork.length)
}
\arguments{
\item{species.code}{3 letter GCMRC species code (i.e., "FMS", "HBC", "RBT")}

\item{fork.length}{(mm)}
}
\value{
weight (g)
}
\description{
Note that weight can vary widely among individuals and over time, predicted weights
calculated by this function provide a mean weight at a given length, but fish
may weight more or less than predicted weights.
}
\examples{
weight_from_fork("BHS", 226)
}
