% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length_conversions.r
\name{weight_from_total}
\alias{weight_from_total}
\title{Predict Weight (g) from Total Length (mm)}
\usage{
weight_from_total(species.code, total.length)
}
\arguments{
\item{species.code}{3 letter GCMRC species code (i.e., "FMS", "HBC", "RBT")}

\item{total.length}{(mm)}
}
\value{
weight (g)
}
\description{
Note that weight can vary widely among individuals and over time, predicted weights
calculated by this function provide a mean weight at a given length, but fish
may weight more or less than predicted weights.
}
\examples{
weight_from_total("RBT", 307)
}
